n8 type
n8 is an immediate type to represent 8 bit value, as in $ff in below

add A, $ff // add A, n8

in emulation this value always come from but with fetch byte function.

the fetch byte function will return a byte which is a type alias for
uint8_t:
add A, [HL]

here [HL] return a byte from the memory at the value of HL. It's
different from n8. This distinction makes implementing instructions more
easier.