Changed ldh [n16], A instruction signature
I aimed to keep the same signature of the instruction and its
implementation but now I realized such a purism is so much for an emu
development which is full of hacks.

Also the Bus::read_write() member is posing problematic because I can't
decide whether it's, well, read or write because it return byte &. I'll try to
reduce calls to it.